+++ app/usecases/commands/registerProduct.test.ts
@@ -18,7 +18,7 @@ const mockTags: ProductTag[] = [
 ]
 
 const orderRepository = {} satisfies Partial<
-  typeof import("../repositories").orderRepository
+  typeof import("../repositories-provider").orderRepository
 >
 
 const productRepository = {
@@ -58,9 +58,9 @@ const productRepository = {
     price: product.price ?? 100,
     stock: product.stock ?? 1,
   })),
-} satisfies Partial<typeof import("../repositories").productRepository>
+} satisfies Partial<typeof import("../repositories-provider").productRepository>
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
