+++ app/usecases/commands/removeOrder.test.ts
@@ -11,13 +11,13 @@ import type { DbClient, TransactionDbClient } from "../../libs/db/client"
 
 const orderRepository = {
   deleteOrder: mock(async () => undefined),
-} satisfies Partial<typeof import("../repositories").orderRepository>
+} satisfies Partial<typeof import("../repositories-provider").orderRepository>
 
 const productRepository = {} satisfies Partial<
-  typeof import("../repositories").productRepository
+  typeof import("../repositories-provider").productRepository
 >
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
