+++ app/usecases/commands/setOrderDetails.test.ts
@@ -24,13 +24,13 @@ const orderRepository = {
         status: order.status ?? "pending",
       }) as Order,
   ),
-} satisfies Partial<typeof import("../repositories").orderRepository>
+} satisfies Partial<typeof import("../repositories-provider").orderRepository>
 
 const productRepository = {} satisfies Partial<
-  typeof import("../repositories").productRepository
+  typeof import("../repositories-provider").productRepository
 >
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
