+++ app/usecases/commands/setOrderStatus.test.ts
@@ -22,13 +22,13 @@ const orderRepository = {
     totalAmount: 1000,
     status: order.status === undefined ? "pending" : order.status,
   })),
-} satisfies Partial<typeof import("../repositories").orderRepository>
+} satisfies Partial<typeof import("../repositories-provider").orderRepository>
 
 const productRepository = {} satisfies Partial<
-  typeof import("../repositories").productRepository
+  typeof import("../repositories-provider").productRepository
 >
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
