+++ app/usecases/queries/exportProductCatalogCsv.test.ts
@@ -16,15 +16,15 @@ const mockProducts: Product[] = []
 const mockTags: ProductTag[] = []
 
 const orderRepository = {} satisfies Partial<
-  typeof import("../repositories").orderRepository
+  typeof import("../repositories-provider").orderRepository
 >
 
 const productRepository = {
   findAllProductsOrderByIdAsc: mock(async (_) => mockProducts),
   findAllProductTagsByIds: mock(async (_) => mockTags),
-} satisfies Partial<typeof import("../repositories").productRepository>
+} satisfies Partial<typeof import("../repositories-provider").productRepository>
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
