+++ app/usecases/queries/getOrderDeletePageData.test.ts
@@ -17,13 +17,13 @@ const mockOrder: Order = {
 
 const orderRepository = {
   findOrderById: mock(async (): Promise<Order | null> => mockOrder),
-} satisfies Partial<typeof import("../repositories").orderRepository>
+} satisfies Partial<typeof import("../repositories-provider").orderRepository>
 
 const productRepository = {} satisfies Partial<
-  typeof import("../repositories").productRepository
+  typeof import("../repositories-provider").productRepository
 >
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
