+++ app/usecases/queries/getOrderProgressManagerComponentData.test.ts
@@ -68,13 +68,13 @@ const orderRepository = {
   findAllOrdersByInactiveStatusOrderByUpdatedAtDesc: mock(
     async () => inactiveOrders,
   ),
-} satisfies Partial<typeof import("../repositories").orderRepository>
+} satisfies Partial<typeof import("../repositories-provider").orderRepository>
 
 const productRepository = {} satisfies Partial<
-  typeof import("../repositories").productRepository
+  typeof import("../repositories-provider").productRepository
 >
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
