+++ app/usecases/queries/getOrderRegistrationFormComponentData.test.ts
@@ -27,15 +27,15 @@ const mockProducts: Product[] = [
 ]
 
 const orderRepository = {} satisfies Partial<
-  typeof import("../repositories").orderRepository
+  typeof import("../repositories-provider").orderRepository
 >
 
 const productRepository = {
   findAllProductsOrderByIdAsc: mock(async () => mockProducts),
   findAllProductTags: mock(async () => mockTags),
-} satisfies Partial<typeof import("../repositories").productRepository>
+} satisfies Partial<typeof import("../repositories-provider").productRepository>
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
