+++ app/usecases/queries/getProductDeletePageData.test.ts
@@ -18,15 +18,15 @@ const mockTags: ProductTag[] = [
 ]
 
 const orderRepository = {} satisfies Partial<
-  typeof import("../repositories").orderRepository
+  typeof import("../repositories-provider").orderRepository
 >
 
 const productRepository = {
   findProductById: mock(async (): Promise<Product | null> => mockProduct),
   findAllProductTagsByIds: mock(async () => mockTags),
-} satisfies Partial<typeof import("../repositories").productRepository>
+} satisfies Partial<typeof import("../repositories-provider").productRepository>
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
