+++ app/usecases/queries/getProductDeletePageData.ts
@@ -2,7 +2,7 @@ import { MAX_TAGS_PER_PRODUCT } from "../../domain/product/constants"
 import type Product from "../../domain/product/entities/product"
 import type ProductTag from "../../domain/product/entities/productTag"
 import type { DbClient } from "../../libs/db/client"
-import { productRepository } from "../repositories"
+import { productRepository } from "../repositories-provider"
 
 const { findProductById, findAllProductTagsByIds } = productRepository
