+++ app/usecases/queries/getProductEditPageData.ts
@@ -1,6 +1,6 @@
 import type Product from "../../domain/product/entities/product"
 import type { DbClient } from "../../libs/db/client"
-import { productRepository } from "../repositories"
+import { productRepository } from "../repositories-provider"
 
 const { findProductById } = productRepository
