+++ app/usecases/queries/getProductRegistrationFormComponentData.ts
@@ -1,7 +1,7 @@
 import { MAX_STORE_PRODUCT_TAG_COUNT } from "../../domain/product/constants"
 import type ProductTag from "../../domain/product/entities/productTag"
 import type { DbClient } from "../../libs/db/client"
-import { productRepository } from "../repositories"
+import { productRepository } from "../repositories-provider"
 
 const { findAllProductTags } = productRepository
