+++ app/usecases/queries/getProductsManagementPageData.test.ts
@@ -33,7 +33,7 @@ const mockProducts: Product[] = [
 ]
 
 const orderRepository = {} satisfies Partial<
-  typeof import("../repositories").orderRepository
+  typeof import("../repositories-provider").orderRepository
 >
 
 const productRepository = {
@@ -45,9 +45,9 @@ const productRepository = {
     { stock: 3 },
   ]),
   findAllProductTagsByIds: mock(async () => mockTags),
-} satisfies Partial<typeof import("../repositories").productRepository>
+} satisfies Partial<typeof import("../repositories-provider").productRepository>
 
-mock.module("../repositories", () => ({
+mock.module("../repositories-provider", () => ({
   orderRepository,
   productRepository,
 }))
